requirements.txt
Flask
tensorflow
scikit-learn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
matplotlib
